New native file access functions.
These functions are bases on the Node.js filesystem API. The following methods are implemented:
  fs.showOpenDialog(allowMultipleSelection, chooseDirectory, title, initialPath, fileTypes, callback)
  fs.readdir(path, callback)
  fs.stat(path, callback)
  fs.readFile(path, [encoding], callback)

(showOpenDialog is not part of the Node API)

All functions require a callback. The callback always has the signature callback(err, data), where err
is one of:

  brackets.fs.NO_ERROR
  brackets.fs.ERR_UNKNOWN
  brackets.fs.ERR_INVALID_PARAMS
  brackets.fs.ERR_NOT_FOUND
  brackets.fs.ERR_CANT_READ
  brackets.fs.ERR_UNSUPPORTED_ENCODING

More errors will be added as more file system functions are added.

Tests are added for the new functions.

This commit also refactors the brackets_extensions JavaScript into an external file stored in the Resources directory.